runnig project! need to do: 1. check open cv (minimal gui mode) in soudry's pc 2. add comments and logger 3. check logger printing 4. check results 5. check TODOs

+++ final/KDArray.h
@@ -31,7 +31,7 @@ int comp_by_coor(const void* a, const void* b);
 
 typedef struct coorToComp {
 	int pointIndex;
-	int coorData;
+	double coorData;
 } CTC;
 void print_KDArr(SPKDArray* arr);
